requirements.txt
PyGithub==1.58.1
supervisely==6.72.43